+++ components/AssistantPanel.tsx
@@ -17,6 +17,7 @@ import { useEditorStore } from "@/hooks/useEditorStore";
 import { useShare } from "@/hooks/useShare";
 import { CapabilityBadge } from "./CapabilityBadge";
 import { AIModeBadge } from "./AIModeBadge";
+import { ChatBrainBadge } from "./ChatBrainBadge";
 import { QuickReplies } from "./QuickReplies";
 import { PlanPreview } from "./PlanPreview";
 import { BriefingCard } from "./BriefingCard";
@@ -203,6 +204,7 @@ export function AssistantPanel({
           </span>
         </div>
         <div style={{ display: "flex", alignItems: "center", gap: 6 }}>
+          <ChatBrainBadge />
           <AIModeBadge />
           <CapabilityBadge />
         </div>